test(TrackedTripTestContest): Move start tracking methods to this class.
